extract sentiment type, add parse func
